# Conflicts:
#	src/main/java/com/yunhuakeji/attendance/biz/impl/UserRoleManageBizImpl.java

+++ src/main/java/com/yunhuakeji/attendance/biz/StudentClockBiz.java
@@ -30,5 +30,5 @@ public interface StudentClockBiz {
 
     Result<Byte> getStudentClockStatusByDay(Long studentId);
 
-    Result<List<TimeClockStatusDTO>> listByWeekNumber(Long studentId, Long weekNumber);
+    Result<List<TimeClockStatusDTO>> listByWeekNumber(Long studentId, int weekNumber);
 }
